Added new timestamps
I periodically check for new entries and while looking at the GitHub repo for the website I noticed that dates had changed on entries. I went through to check them all and it appears that all entries have had their dates changed, except for the placeholders. Whereas before they were marked December 31st, they now have a variety of dates between January 26th and February 3rd 2016. It is still unclear what the date signifies, though based on Nick's story it is not the date of the incident described in the entry because his occurred many years ago. It could be the date of submission, though in Nick's case that means he submitted his story over four months prior to when he wrote it. I think it should also be noticed that his entry is dated the earliest and he's the only person on the website who has a clear identity.