Reordered entries based on dates
I re-ordered the presentation of the entries to reflect a timeline based on their dates. Whatever the dates mean, they are on a one-entry-per-day basis between January 25th and February 3rd. Since "Alice" and "Briar" are still left, I'm going to predict that those entries will go on January 29th and 30th because those dates are missing from the timeline.