Noticed different color of text
I had not paid close attention to the way the text of the GitHub PNG images of the entries was presented, but I just noticed that it is blue in odd places not consistent with, perhaps, highlighting, italics, or bold. Since the website has already confessed to be editing entries for the sake of hiding identifying details, it is possible these are tamper marks. Some entries have very little, like Daniel's https://archive.is/hVV6V, while others have a lot, like Phoenix's https://archive.is/jUxNM, where whole paragraphs are coloured blue.