Added JTRIG and CIJ speech
Necessary disclosure: While digging through Appelbaum's tweets surrounding the suspension in March 2015, I found two tweets made around the same day as when the letter from Leckrone was sent, which may have reflected his reaction to it. The JTRIG one stood out to me considering the nature of what he had been accused of a few days before. I've had a few requests (or rather questions of why I haven't done this yet) to add discussions about the possibility of state actors being involved, and so far I hadn't found an opportune place to even mention it considering how early it still is in this case. However I do think adding it is justified, because even those who don't believe this is being orchestrated by "state actors" understand that Appelbaum's work has certainly been 'disruptive' enough to attract that kind of attention. I think it is only fair, in terms of offering each side the opportunity to make their case, that proper context is given to all of this. Remember that Tor has not endorsed or even mentioned the victims' website officially; even if it is run by members of that organization, it is still as of now a very non-transparent operation that has so far acted with questionable ethics towards both the accusers and the accused. It does not escape my notice that many of the same people who have been warning us about these kinds of operations for years are now turning around to say that those who believe or are wary of them are "paranoid," which I have no doubt was the cruel label they were marked with in the years leading up to Snowden. I think it is dishonest to try to prove how great an individual's power was in a movement while simultaneously trying to downplay it in the context of their vulnerability to such hidden operations. This all appears to have been building for a long time - why it has all come to a head now should be explored. Regardless, proving whether JTRIG was involved is not the reason I'm including this information, but rather to pre-emptively include it in anticipation of it being used as an argument by Appelbaum's defense (I would be surprised of he didn't). As always, you may see your own patterns and draw your own conclusions.