Updated section on Twitter account
The Twitter account which linked to Caleb's repository yesterday is now linking to mine after I suggested to Caleb he add a disclaimer, despite the fact that I've blocked them. I had anticipated this was going to happen at some point, so I updated my disclaimer yesterday when I saw the account. I have one Twitter account (@J9Roem) and that is the only one which represents me or this repository. Anyone can link to my repository, as I've been cataloguing with the media mentions; I have no control over this. I don't care if the mention is positive or negative -- if someone is violating my anti-harassment policy then it is not appreciated. This kind of activity does no one, especially victims, any good. This kind of behaviour only rewards the troll.