Updated story of "River"
I haven't had anyone take issue with how I've summarised the victim stories and I've tried to be conscious about how I classify the different allegations, which is difficult when the line between what people call "sexual assault" vs. "rape" varies. It also varies according to whether they're addressing the legal definitions or not (since many people in this space have publicly stated they don't care about law, I think it's intellectually dishonest to then _only_ cite German law or American law etc. as your source for arguing about the distinction between the two). My understanding has been that any non-consensual (that is, whether the victim says some form of "no," resists, or for some reason is incapable of doing either) interaction, with physical contact that is of a sexual nature (involving the sexual organs of either person) is sexual assault, and that rape is a more serious subset of sexual assault where penetration occurs (with sexual organs or foreign objects). I automatically classify any 1) non-consensual interaction of 2) a sexual nature that 3) involves penetrative intercourse as rape. There have been many debates about whether sexual assault without penetration + unconscious or intoxicated state also 'deserves' being called "rape," but I think it's fair to say that they are both horrible and that the 'correct' use of this language is still ambiguous even in a single culture or country. What also makes it difficult is that the language of many of the stories is rather vague and, in the case of Forest/Lovecruft & "River," they either call it "sexual assault" themselves or do not directly state it/ do not given enough detail to tell the difference.

As I was updating this investigation with Lovecruft's account of various events, I also saw this comment on Reddit https://www.reddit.com/r/TOR/comments/4nzq91/megathread_jacob_appelbaum/d49gxk0 (in case it disappears: https://archive.is/8wSZc) which not only accused me of "glossing over a pattern of behaviour by Appelbaum" but that I "even [chose] to downplay the public rape of an unconscious woman by calling it 'sexual assault'." It was not received positively but considering that I anticipate getting accused of this again, I'm going to respond to it.

(First of all, the comment calls me an "it"; thanks for the ironic objectification). Considering that my 'Allegations' section is the largest section in this investigation and I have documented and clearly sourced all those allegations against Appelbaum, whether by anonymous or known claimants, whether it's an actual criminal allegation or a case of employee misconduct, where for many I summarized and say "also" (which obviously implies that someone else has made a similar allegation and therefore does represent of a "pattern of behaviour"), I would like to know how exactly I "glossed over" anything. I have yet to see any other 'official' journalists who are reporting on this story be as thorough as I have, but if my 'glossing' is so egregious that you think it's worthy of complaining about, please link me to someone else who is being more comprehensive than I am and I will emulate them. Also, I am doing this out of intrinsic motivation and not receiving any financial compensation for the work I am doing in this investigation -- so please, let me know if I'm just wasting my time.

Your second allegation is that I "[chose] to downplay" what happened in the story of "River," and this allegation is far more insulting than the first. To begin with, I think any gradient of sexual assault is a serious matter; I don't need the further classification of "rape" to think that, but maybe that's just me. Secondly, the reason I classified it as "sexual assault" was because of the language used in the story, which comprised a variation of non-consensual sexual contact: "going down on me," "touching," and "having sex with me." "Going down on" someone without their consent would be sexual assault but that phrase has been used in a variety of ways, not all of which include penetrative intercourse (ex. grinding, 'dry humping'). "Touching" is also non-specific and could be anything from shoulder-rubbing to fingering (which could be penetrative). Finally, "having sex with me," is much more likely to be describing penetration but is still non-specific when you consider that all of the actions described in the story were of a sexual nature and obviously a lot of sex isn't about penetration.

It wasn't until Lovecruft's blog post that there was a clear allegation not only of penetration being involved, but that the "friend of Appelbaum" had also raped her, when "River" only used the vague term of "touching." I have since added the details Lovecruft provided, I am keeping the original label of "sexual assault" not to downplay what allegedly happened but to reflect the vagueness of the language used, which was then updated by Lovecruft's post and maybe will be clarified further over the course of the investigation.

Having said all this, I don't think I "[chose] to downplay" anything, because that assumes I made light of a situation despite more information. Furthermore, if you think initially describing the story as "sexual assault," even given the language ambiguity, is a 'downplay,' perhaps you need to re-evaluate how you respond to the term "sexual assault." "Assault" on its own is a varied but serious allegation which can include striking someone that doesn't even leave a bruise to beating someone to near-death; there is nothing light about it. Sexual assault is not only serious but can also leave long-term emotional and psychological marks even if there is no physical marks of any kind. Again, there is nothing light about it.

If, after reading this explanation, anyone still thinks I "glossed over" or "downplayed" something, I respond to pull-requests, Issues, emails, tweets etc. and to date I have acknowledged and answered all of them, so you shouldn't have a problem coming forward and trying to correct me.