Replaced video link with cached version
The Korean tour video has been removed from Vimeo (I don't know the reason, it could have been copyright infringement). I made a local copy when I found it just in case this happened; for now I've replaced the video link with the cached version to prove the page was there, but I will try to get a new live link for the video soon.