New subheadings, more organisation /1
This is the first of (what will probably be) two updates to add new subheadings to make it easier to navigate under the larger headings. Nothing was removed but a few paragraphs were moved for organisation or chronology reasons. These changes were made based on: https://github.com/Enegnei/JacobAppelbaumLeavesTor/pull/9