Removed repetition of Shepard's full name
In case it wasn't noticed yet, my strategy for addressing people in this investigation is to use their full name (with links on further identity information) only once when referring to them for the first time. All further references after that only use their surname, unless there are two people with the same surname.