Removed last entry until clarification
I removed the entry I made from Golem.de as I'm not entirely sure how to interpret "offene Rechnungen" because it can mean two different things depending on if it's meant literally ("unpaid / outstanding invoices") or as a figure of speech (like "alte Rechnungen begleichen," meaning 'to settle old scores / grievances'). Contextually it is probably more likely to mean the latter but this usage is not familiar to me, so I have contacted the journalist for clarification. If I was correct the first time, I will re-add it.