Linked to document cloud, improve navigation
I previously linked to the Internet Archive mirror of the Appelbaum Gmail case repeatedly for each citation, but that's rather unnecessary as it does not really help readers to find where the quote comes from. To improve navigation I've linked to the document cloud instead since I can link directly to the pages I need, and even if that is taken down (which is very low), the whole set is still archived anyway.