Updated "latest updates" section
Note on "latest updates" section - there may be times where, after I've updated a section and noted it here, I will subsequently update it again in the following days. I will not re-add an update, so if an update note is still included there that means I've added even more since you saw it the first time. This section acts as an indicator of which sections are being actively edited, so even if you've already checked a note once, you may have to check it again.